Fix uint16_t

+++ application_processor/src/board_link.c
@@ -48,7 +48,7 @@ i2c_addr_t component_id_to_i2c_addr(uint32_t component_id) {
  *
  * Function sends an arbitrary packet over i2c to a specified component
 */
-int send_packet(i2c_addr_t address, uint16_t len, uint8_t* packet) {
+int send_packet(i2c_addr_t address, uint8_t len, uint8_t* packet) {
 
     int result;
     result = i2c_simple_write_receive_len(address, len);
@@ -76,7 +76,7 @@ int send_packet(i2c_addr_t address, uint16_t len, uint8_t* packet) {
  * @return int: size of data received, ERROR_RETURN if error
 */
 int poll_and_receive_packet(i2c_addr_t address, uint8_t* packet) {
-
+    
     int result = SUCCESS_RETURN;
     while (true) {
         result = i2c_simple_read_transmit_done(address);
